Fix: End Screen Bug with Character Change (#60)
Addressing a bug where players could exit a match prior to the end screen and alter their character, resulting in unintended consequences for the remaining players' end screen experience. Normally, fetching new data is possible, but in this instance, it is not. This issue specifically arises when a player utilizes a GG Boost during a match while another player has left and subsequently changed their character, banner, title, or other related elements.